requirements.txt
asgiref==3.9.1
Django==5.2.4
django-allauth==65.9.0
django-oauth==1.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
